requirements.txt
mkdocs-material
mkdocs-minify-plugin
mkdocs-redirects
mkdocs-macros-plugin
./plugins/mkdocs-redirects/
